Make sure push_read_back results in the event being ready.
If you push back read data from a socket and it is subsequently idle,
you need to push the data back to a buffer and mark the event as having
data ready to read somehow.  There is already a facility for Danga's
PostEventLoop() function to be called, which is supposed to clear these
synthetic read events.  However, a bit wasn't set on the tasks' list of
interesting events, and this caused the pushed back data not to 'wake
up' the task and continue.  So, set the bit in the right place.