Wake and sleep events with "push back" data
This module's watch_read function, which is monkeypatched into
Danga::Socket, sets up an AnyEvent io event watcher on reads; but in
the case where synthetic reads ("push_back_read") are in effect, this
event will never fire.  This case can be detected, and the bit flipped
as in a previous change.  It also has a converse, which is to clear it
in the same situations.